// electron/storage/types/IStorage.ts
import Entity from './Entity'
import SortType from './SortType'

interface IStorage<T extends Entity> {
  list: (searchTerm: string, sort: SortType) => Promise<T[]>
  get: (identity: string) => Promise<T>
  add: (entity: Omit<T, '_id'>) => Promise<T>
  update: (
    identity: string,
    fieldsToUpdate: Partial<Omit<T, '_id'>>
  ) => Promise<T>
  delete: (identity: string) => Promise<void>
  find: (entity: Partial<Omit<T, '_id'>>) => Promise<T>
}

export default IStorage